+++ src/components/SukohForm/index.js
@@ -135,6 +135,7 @@ export class SukohForm extends React.Component<SukohFormProps, SukohFormState>{
                     pageUrl={this.props.pageUrl}
                     values={this.props.values}
                     onChange={this.handleFormChange.bind(this)}
+                    onOpenInEditor={this.props.onOpenInEditor}
                     />
                 <FloatingActionButton
                     style={{