+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # PoppyGo App CHANGELOG
 
-
 ## vNEXT
 
 ## v0.6.6
 
+- improve bundle-manager and image thumb layout
+- open single item in editor
+- open collection item in editor
 - don't bother users with valid keys, let them enter titles and auto generate key
 - delete directory too when deleting a pageBundle
 - open entry after creation collections